requirements.txt
aerich==0.5.7
aiogram==2.25.1
aiohttp==3.8.1
aioredis==2.0.0
aiosqlite==0.18.0
async-timeout==4.0.0
asyncpg==0.24.0
attrs==21.2.0
Babel==2.9.1
chardet==4.0.0
click==8.0.1
colorama==0.4.4
envparse==0.2.0
hiredis==2.0.0
idna==3.3
iso8601==0.1.16
loguru==0.5.3
multidict==5.2.0
pydantic==1.8.2
PyPika==0.48.2
pytz==2021.3
tortoise-orm==0.16.17
typing-extensions==4.7.1
win32-setctime==1.0.3
yarl==1.7.0
python-docx==0.8.11
lxml==4.9.3
google-auth==1.35.0  
google-auth-oauthlib==0.4.6  
google-api-python-client==2.22.0  
